Split CardList.dart into smaller files
